requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.7
cryptography==37.0.1
diffusers==0.9.0
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.8.0
Flask==3.0.1
flit_core==3.6.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.11.1
idna==3.3
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.2.1
openai==1.10.0
opencv-python==4.8.0.76
packaging==21.3
Pillow==9.2.0
pip==21.2.4
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pycosat==0.6.3
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.1
regex==2022.10.31
requests==2.27.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sympy==1.11.1
tiktoken==0.5.2
toolz==0.12.0
tqdm==4.64.0
typing_extensions==4.9.0
urllib3==1.26.9
Werkzeug==3.0.1
wheel==0.37.1
zipp==3.11.0

